CLAP-70 Feat: processorOrder 자료형 int -> long 변경
<footer>
- 관련: #29

+++ src/main/java/clap/server/adapter/outbound/persistense/entity/task/TaskEntity.java
@@ -43,7 +43,7 @@ public class TaskEntity extends BaseTimeEntity {
     private TaskStatus taskStatus;
 
     @Column(nullable = false)
-    private int processorOrder;
+    private long processorOrder;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "reviewer_id")
